Moving Functions (#40)
* Moving Functions

Here I move around some functions and inputs.

All of the functions  from "Get_Fiber Response" and "Get_Rendering_Animations" has been changed.

This breaks things! Some inpurts are now missing. I will fix it as I go through my final cleaning making the test suite the test functions.

* Fixing Imports, updated fibre functions

Changed imports to work correctly
General bug fixing on plot functions.
Changed the fiber data to save to OBD
Made savefiberdata2D save to OBD by default, made function availible to users.

* Moved Save Fiber Data
